Use the new allocator

+++ agb/src/display/tiled/mod.rs
@@ -18,7 +18,7 @@ impl Tile {
     }
 
     fn tile_index(self) -> TileIndex {
-        TileIndex::new(self.0 & ((1 << 10) - 1))
+        TileIndex::new(self.0 as usize & ((1 << 10) - 1))
     }
 }
